Refactor into new student flow into newflow module
