+++ app/views/newflow/login_signup/_back_arrow.html.erb
@@ -1,0 +1,3 @@
+<%= link_to path, class: 'back-arrow' do%>
+    <i class="fa fa-arrow-left"></i>
+<% end %>
