+++ app/views/newflow/login_signup/_tab_group.html.erb
@@ -2,7 +2,7 @@
     <%= link_to I18n.t(:"login_signup_form.log_in"), newflow_login_path,
                     class: "tab login #{active_class}"
     %>
-    <%= link_to I18n.t(:"login_signup_form.sign_up"), newflow_welcome_path,
+    <%= link_to I18n.t(:"login_signup_form.sign_up"), newflow_signup_path,
                     class: "tab signup #{active_class}"
     %>
     <a class="tab tab--placeholder"></a>
