+++ app/views/newflow/login_signup/confirm_social_info_form.html.erb
@@ -23,7 +23,7 @@
 
                 <div class="content role-section">
                     <div>
-                        <a href=<%= newflow_signup_path %> class="back-arrow"><i class="fa fa-arrow-left"></i></a>
+                        <%= render partial: 'back_arrow', locals: { path: newflow_signup_path } %>
                         <div class="role-button"><%= I18n.t(:"login_signup_form.student") %></div>
                     </div>
                 </div>
