+++ app/views/newflow/login_signup/social_login_failed.html.erb
