+++ app/views/newflow/login_signup/welcome.html.erb
@@ -15,7 +15,7 @@
                     <div class="join-as">
 
                         <div>
-                            <%= link_to newflow_signup_path do%>
+                            <%= link_to newflow_signup_student_path do%>
                                 <div class="join-as__role student">
                                     <span><%= I18n.t(:"login_signup_form.student") %></span>
                                     <%= image_tag('student-icon.svg') %>
