--- spec/controllers/login_signup_controller_spec.rb
@@ -1,5 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe LoginSignupController, type: :controller do
-
-end
