+++ spec/controllers/newflow/login_signup_controller_spec.rb
@@ -1,0 +1,7 @@
+require 'rails_helper'
+
+module Newflow
+  RSpec.describe LoginSignupController, type: :controller do
+
+  end
+end
