+++ spec/features/newflow/student_signup_flow.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+feature 'Student signup flow', js: true do
+
+end
